should actually change price while buying aswell

+++ src/server/game/Handlers/ItemHandler.cpp
@@ -660,11 +660,11 @@ void WorldSession::SendListInventory(ObjectGuid vendorGuid)
                 continue;
             }
 
-            float price = 0;
+            uint64 price = 0;
             if (!vendorItem->ExtendedCost)
             {
-                price = floor(itemTemplate->GetBuyPrice() * discountMod);
-                price = itemTemplate->GetBuyPrice() > 0 ? std::max(1.00f, price) : price;
+                price = uint64(floor(itemTemplate->GetBuyPrice() * discountMod));
+                price = itemTemplate->GetBuyPrice() > 0 ? std::max(1u, price) : price;
             }
 
             if (int32 priceMod = _player->GetTotalAuraModifier(SPELL_AURA_MOD_VENDOR_ITEMS_PRICES))
@@ -676,7 +676,7 @@ void WorldSession::SendListInventory(ObjectGuid vendorGuid)
             item.Type = vendorItem->Type;
             item.Quantity = leftInStock;
             item.StackCount = itemTemplate->GetBuyCount();
-            item.Price = uint32(price);
+            item.Price = price;
             item.DoNotFilterOnVendor = vendorItem->IgnoreFiltering;
             item.Refundable = itemTemplate->HasFlag(ITEM_FLAG_ITEM_PURCHASE_RECORD) && vendorItem->ExtendedCost && itemTemplate->GetMaxStackSize() == 1;
